% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_args.r
\name{.mcmc}
\alias{.mcmc}
\title{Shared argument(s)}
\arguments{
\item{mcmc}{An object of class \code{mcmc} or \code{mcmc.list}, \emph{or} a \code{list}. If a \code{list}, the function searches down the first element to see if it can find an \code{mcmc} or \code{mcmc.list} object, then uses this if it can.}
}
\description{
Shared argument(s)
}
