% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emc_trace_dens.r
\name{emc_trace_dens}
\alias{emc_trace_dens}
\title{Trace plots and density plots}
\usage{
emc_trace_dens(
  x,
  var = NULL,
  i = NULL,
  j = NULL,
  k = NULL,
  l = NULL,
  trace = TRUE,
  loess = TRUE,
  nrows = 5,
  file = NULL,
  ...
)
}
\arguments{
\item{x}{An object of class \code{mcmc} or \code{mcmc.list}, or a \code{list}. If a \code{list}, the function searches down the first element to see if it can find an \code{mcmc} or \code{mcmc.list} object, then plots this if it can.}

\item{var}{Name of the variable(s). The outcome depends on the definitions of \code{i}, \code{j}, \code{k}, and \code{l}. Please see the help for \code{\link{emc_vars}} for more explanation on how to specify this argument, plus \code{i}, \code{j}, \code{k}, and \code{l}.}

\item{i, j, k, l}{Indices used to specify variable names.}

\item{trace}{Logical. If \code{TRUE} (default), generate a trace plot.}

\item{loess}{Logical. If \code{TRUE} (default), draw a LOESS line for each chain in a trace plot.}

\item{nrows}{Number of rows of graphs to display per "page".}

\item{file}{Either \code{NULL} (default) or name of a file to which to save. Specifying a PDF file is especially helpful for cases where the}

\item{...}{Arguments to pass to \code{\link[ggplot2]{ggsave}}.}

\item{density}{Logical. If \code{TRUE} (default), generate a density plot.}
}
\value{
A \pkg{ggplot2} \code{ggplot} graphic object and (possibly) a file saved.
}
\description{
Trace plots and density plots
}
