% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emc_vars.r
\name{emc_vars}
\alias{emc_vars}
\title{Match variable names to results from an MCMC}
\usage{
emc_vars(var, i = NULL, j = NULL, k = NULL, l = NULL, x = NULL)
}
\arguments{
\item{var}{Name of the variable(s). The outcome depends on the definitions of \code{i}, \code{j}, \code{k}, and \code{l}.
\itemize{
        \item \code{var = NULL}: Plot all monitored parameters.
        \item \code{var =} a character vector and \code{i}, \code{j}, and \code{k} are all \code{NULL}: Plot the variable(s) exactly as named.
        \item \code{var =} a character vector and \code{i} is a numeric vector: Plot all parameters with the pattern "\code{var*}" where \code{*} is \code{i}. For example: \code{beta1}, \code{beta2}, and \code{beta3}.
        \item \code{var =} a character vector and \code{i} is \code{TRUE}: Plot all parameters with the pattern \code{beta*} where \code{*} is 1 through the maximum value found in the \code{mcmc} object.
    \item \code{var =} a character vector and \code{j} is a numeric vector: Plot all parameters with the pattern "\code{beta[j]}".
    \item \code{var =} a character vector and \code{j} and \code{k} are a numeric vectors: Plot all parameters with the pattern "\code{beta[j, k]}".
    \item \code{var =} a character vector and \code{j}, \code{k}, and \code{l} are a numeric vectors: Plot all parameters with the pattern "\code{beta[j, k, l]}".
    \item \code{var =} a character vector and \code{i}, \code{j}, \code{k}, and \code{l} are a numeric vectors: Plot all parameters with the pattern "\code{beta*[j, k, l]}" where \code{*} is \code{i}.
    \item \code{var =} a character vector and any of \code{i}, \code{j}, \code{k}, and/or \code{l} are \code{TRUE}: Plot the given pattern up the the maximum value of \code{i}, \code{j}, \code{k}, and/or \code{l}.
}}

\item{x}{Either \code{NULL} (default), or an object of class \code{mcmc} or \code{mcmc.list}, or a \code{list}. If \code{NULL}, then the returned variable names are not necessarily found in the MCMC object.  If a \code{list}, the function searches iteratively down the first element of the list to see if it can find an \code{mcmc} or \code{mcmc.list} object.}
}
\value{
Character vector of variables.

var <- 'beta'
emc_vars(var)
emc_vars(var, i=0:1)
emc_vars(var, j=1:2)
emc_vars(var, i=0:1, j=1:2)
emc_vars(var, j=1:2, k=1:3)
emc_vars(var, i=0:1, j=1:2, k=1:2)
emc_vars(var, j=1:2, k=1:2, l=1:2)
emc_vars(var, i=0:1, j=1:2, k=1:2, l=1:2)
}
\description{
This function automates finding of indexed variable in an MCMC object. These include, for example, variables like \code{beta1}, \code{beta2}, \code{beta3}, etc. or variables like \code{beta[1]}, \code{beta[2]}, \code{beta[1]}, etc. Only variable named in the MCMC object are returned.
}
