% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emcemc.r
\docType{package}
\name{emcemc}
\alias{emcemc}
\title{emcemc: Analysis of MCMC chains}
\description{
Helper functions for analyzing MCMC chains. Includes trace plots and density plots. This package automates extraction of specific parameters, including indexed coefficients such as "beta[1, 1]", "beta[1, 2]", "beta[1, 3]", and similar.
}
\details{
Create an issue on \href{https://github.com/adamlilith/enmemc/issues}{GitHub}.
}
\section{MCMC chain manipulation}{

		\code{\link{emc_stack_chains}} "Stack" multiple MCMC chains on one another \cr
}

\section{MCMC diagnostics}{

		\code{\link{emc_trace_dens}} Trace plots and density plots \cr
}

\author{
Adam B. Smith
}
\keyword{internal}
