% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hammer_cbind.r
\name{hammer_cbind}
\alias{hammer_cbind}
\title{Combine multiple MCMC chains with different monitored variables}
\usage{
hammer_cbind(..., summarize = TRUE)
}
\arguments{
\item{...}{Two or more objects of class \code{mcmc} or \verb{mcmc.list``, or a }list\verb{with the first element being an}mcmc.list\verb{. All inputs must be of the same class (e.g., you can't combine an }mcmc\verb{object with an}mcmc.list`).}

\item{summarize}{Logical: If \code{TRUE} (default), and the input is an \code{mcmc.list} or list of lists, then calculate summary statistics.}
}
\value{
An \code{mcmc} object, an \code{mcmc.list}, or a \code{list}.
}
\description{
This function takes an \code{mcmc.list} and \code{cbind()}'s' them onto one another to create an \code{mcmc.list} that has all the columns of the inputs. It also combines the summary statistics.
}
\examples{

data(mcmc)
cbinded <- hammer_cbind(mcmc)
head(cbinded)
dim(cbinded)

}
