% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hammer_combine.r
\name{hammer_combine}
\alias{hammer_combine}
\title{Combine MCMC chains}
\usage{
hammer_combine(...)
}
\arguments{
\item{...}{Two or more \code{list}s with \code{mcmc.list} objects therein, or a \code{list} with an \code{mcmc} object (i.e., a single chain).}
}
\value{
A \code{list} with the following structure:

\if{html}{\out{<div class="sourceCode">}}\preformatted{.$samples$chain1
.$samples$chain2
.$samples$chain3
.$samples$chain4
(etc.)
.$summary$chain1
.$summary$chain2
.$summary$chain3
.$summary$chain4
(etc.)
.$summary$all.chains
}\if{html}{\out{</div>}}

Chains in the input will be renumbered as needed in the output.
}
\description{
Sometimes, you need to run different MCMC chains in different instances of \strong{R}, then want to combine them. This function does that by creating a \code{list}, typical of outputs from \strong{JAGS}, \strong{nimble}, \strong{stan}, and other MCMC sampler software.
}
\examples{

data(mcmc)
mcmc2 <- mcmc # make a copy so we can combine them
combo <- hammer_combine(mcmc, mcmc2)
str(combo, 2)

}
