% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hammer_extract.r
\name{hammer_extract}
\alias{hammer_extract}
\title{Extract posterior mean, median, or quantiles from an mcmc or mcmc.list}
\usage{
hammer_extract(
  mcmc,
  param,
  i = NULL,
  j = NULL,
  k = NULL,
  l = NULL,
  stat = "mean"
)
}
\arguments{
\item{mcmc}{An object of class \code{mcmc} or \code{mcmc.list}, \emph{or} a \code{list}. If a \code{list}, the function searches down the first element to see if it can find an \code{mcmc} or \code{mcmc.list} object, then uses this if it can.}

\item{param}{Character vector: Name of the variable(s).}

\item{i, j, k, l}{Indices used to specify variable names:
\itemize{
\item \code{i} refers to indices that occur immediately after the variable's "base" name. For example, if \code{param} is \code{beta}, and \verb{i = 0:2', then the function will return }beta0\verb{, }beta1\verb{, and }beta2`.
\item \code{j} refers to the \emph{first} index in square brackets. For example, if \code{param} is \code{beta} and \code{j = 1:3}, then the output will be \code{beta[1]}, \code{beta[2]}, and \code{beta[3]}.
\item \code{k} refers to the \emph{second} index in square brackets. If \code{k} is not \code{NULL}, then \code{j} cannot be \code{NULL}. For example, if \code{param} is \code{beta}, \code{j = 1:2}, and \code{k = 1:3}, then the output will be like \code{beta[1, 1]}, \code{beta[2, 1]}, \code{beta[1, 2]}, \code{beta[2, 2]}, \code{beta[1, 3]}, and \code{beta[2, 3]}.
\item \code{l} refers to the \emph{third} index in square brackets. You must also define \code{j} and \code{k} is you provide \code{l}. For example, if \code{param} is \code{beta}, \code{j = 1:2}, \code{k = 4:5}, and \code{l = 9:10}, then the output will be like \code{beta[1, 4, 9]}, \code{beta[2, 4, 9]}, \code{beta[1, 5, 9]}, \code{beta[2, 5, 9]}, \code{beta[1, 4, 10]}, \code{beta[2, 4, 10]}, and so on.
\item You can also define \code{i} when you define \code{j} alone, \code{j} and \code{k}, or \code{j}, \code{k}, and \code{l} to get outputs like \code{beta0[1]}\verb{, or like \code{beta0[1, 1]}, or like \code{beta0[1, 2, 3]}}.
You must specify all of the indices that would appear in square brackets. For example, for the pattern \code{param[j, k, l]}, you need to define \code{j}, \code{k}, and \code{l}. You can't just define \code{k = 1:2}. You need to set \code{j} and \code{l} equal to numeric/integer vectors and/or logical values.
}}

\item{stat}{Character: Indicates which statistic to extract. Options are \code{'mean'} (default), \code{'median'}, \code{'sd'}, \code{'lower'}, '\verb{upper'}. The \code{lower} and \code{upper} options extract the lower and upper 2.5th and 97.5th quantiles, respectively. Partial matching is allowed.}
}
\value{
An \code{mcmc.list} object or a \code{list}.
}
\description{
Extracts posterior mean, median, standard deviation, or quantiles from an \code{mcmc} or \code{mcmc.list} object.
}
\examples{

data(mcmc)
hammer_extract(mcmc, param = 'beta', j = 1:3)
hammer_extract(mcmc, param = 'beta', j = 1:3, stat = 'median')
hammer_extract(mcmc, param = 'z_hat', j = TRUE, k = TRUE)

}
