% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hammer_head.r
\name{hammer_head}
\alias{hammer_head}
\title{Display the top or bottom rows of an MCMC chain or its summary.}
\usage{
hammer_head(x, param = NULL, i = NULL, j = NULL, k = NULL, l = NULL, n = 6)
}
\arguments{
\item{x}{An \code{mcmc} or \code{mcmc.list} object.}

\item{param}{Character vector: Characters to keep or discard.}

\item{i, j, k, l}{Indices used to specify variable names. Please see \code{\link[=mh_param]{mh_param()}}.}

\item{n}{Number of rows to display. Default is 6.}

\item{chain}{Any of:
\itemize{
\item \code{NULL} (default): Display head/tail of each chain.
\item A numeric vector with integers: Display head/tail of these chains.
}}
}
\value{
A \code{list} object with the head or tail of each chain.
}
\description{
Display the topmost or bottommost rows of an MCMC object.
}
\examples{

data(mcmc)
hammer_head(mcmc, 1)
hammer_tail(mcmc, 1)

}
