% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hammer_rbind.r
\name{hammer_rbind}
\alias{hammer_rbind}
\title{"Stack" multiple MCMC chains on one another}
\usage{
hammer_rbind(mcmc, chains = FALSE)
}
\arguments{
\item{mcmc}{An object of class \code{mcmc} or \code{mcmc.list}, \emph{or} a \code{list}. If a \code{list}, the function searches down the first element to see if it can find an \code{mcmc} or \code{mcmc.list} object, then uses this if it can.}

\item{chains}{Logical: If \code{TRUE}, then the output will have a column named \code{chain} with the chain number of each row.}
}
\value{
Matrix.
}
\description{
This function takes an \code{mcmc.list} and \code{rbind()}'s' them onto one another to create a single, large matrix.
}
\examples{

data(mcmc)
rbinded <- hammer_rbind(mcmc)
head(rbinded)
dim(rbinded)

}
