% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hammer_resummarize.r
\name{hammer_resummarize}
\alias{hammer_resummarize}
\title{Re-calculate summaries of MCMC chains}
\usage{
hammer_resummarize(mcmc, na.rm = FALSE)
}
\arguments{
\item{mcmc}{An \code{mcmc} or \code{mcmc.list} object, or a list of such objects.}

\item{na.rm}{Logical: If \code{FALSE} (default), then summary statistics will be \code{NA} if any value in any iteration is \code{NA} (and it probably should be \code{NA}). However, if you set this to \code{TRUE}, then \code{NA} values are ignored.}
}
\value{
A \code{list} with the following structure:

\if{html}{\out{<div class="sourceCode">}}\preformatted{.$summary$chain1 <matrix>
.$summary$chain2 <matrix>
.$summary$chain3 <matrix>
.$summary$chain4 <matrix>
(etc.)
.$summary$all.chains <matrix>
}\if{html}{\out{</div>}}
}
\description{
This function re-calculates the summary statistics of MCMC chains.
}
\examples{

data(mcmc)
summ <- hammer_resummarize(mcmc)
str(summ)

}
