% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hammer_resummary.r
\name{hammer_resummary}
\alias{hammer_resummary}
\title{Re-calculate summaries of MCMC chains}
\usage{
hammer_resummary(mcmc)
}
\arguments{
\item{mcmc}{An \code{mcmc} or \code{mcmc.list} object, or a list of such objects.}
}
\value{
A \code{list} with the following structure:

\if{html}{\out{<div class="sourceCode">}}\preformatted{.$summary$chain1 <matrix>
.$summary$chain2 <matrix>
.$summary$chain3 <matrix>
.$summary$chain4 <matrix>
(etc.)
.$summary$all.chains <matrix>
}\if{html}{\out{</div>}}
}
\description{
This function re-calculates the summary statistics of MCMC chains.
}
\examples{

data(mcmc)
summ <- hammer_resummary(mcmc)
str(summ)

}
