% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hammer_samples_summaries_summary.r
\name{hammer_samples}
\alias{hammer_samples}
\alias{hammer_summaries}
\alias{hammer_summary}
\title{Extracts a "samples" mcmc.list from a list, or the "summary" of an mcmc.list}
\usage{
hammer_samples(mcmc, fail = TRUE)

hammer_summaries(mcmc, fail = TRUE)

hammer_summary(mcmc, fail = TRUE)
}
\arguments{
\item{mcmc}{A list object, or an \code{mcmc} or \code{mcmc.list} object.}

\item{fail}{Logical: If \code{TRUE} (default), and the object being searched for cannot be found, then fail with an error. If \code{FALSE}, then return \code{NULL}.}
}
\value{
An \code{mcmc.list} object or a \code{list}.
}
\description{
These functions extract the \code{samples} or \code{summary} part of a list object:
\itemize{
\item \code{hammer_samples()}: Extracts a list of MCMC samples.
\item \code{hammer_summaries()}: Extracts a list of MCMC summary matrices.
\item \code{hammer_summary()}: Extracts the "all-chains" summary matrix.
}
}
\section{Functions}{
\itemize{
\item \code{hammer_summaries()}: Extracts a "samples" list from a list

\item \code{hammer_summary()}: Extracts a grand "summary" matrix from a list

}}
\examples{

data(mcmc)
str(hammer_samples(mcmc))
str(hammer_summaries(mcmc))
hammer_summary(mcmc)

}
