% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_combine.r
\name{mc_combine}
\alias{mc_combine}
\title{Combine MCMC chains}
\usage{
mc_combine(..., na.rm = FALSE)
}
\arguments{
\item{...}{Two or more \code{list}s with \code{mcmc.list} objects therein, or a \code{list} with an \code{mcmc} object (i.e., a single chain).}

\item{na.rm}{Logical: If \code{FALSE} (default), then summary statistics will be \code{NA} if any value in any iteration is \code{NA} (and it probably should be \code{NA}). However, if you set this to \code{TRUE}, then \code{NA} values are ignored.}
}
\value{
A \code{list} with the following structure:

\if{html}{\out{<div class="sourceCode">}}\preformatted{.$samples$chain1
.$samples$chain2
.$samples$chain3
.$samples$chain4
(etc.)
.$summary$chain1
.$summary$chain2
.$summary$chain3
.$summary$chain4
(etc.)
.$summary$all.chains
}\if{html}{\out{</div>}}

Chains in the input will be renumbered as needed in the output.
}
\description{
Sometimes, you need to run different MCMC chains in different instances of \strong{R}, then want to combine them. This function does that by creating a \code{list}, typical of outputs from \strong{JAGS}, \strong{nimble}, \strong{stan}, and other MCMC sampler software.
}
\examples{

data(mcmc)
combo <- mc_combine(mcmc, mcmc) # combine "mcmc" with itself
str(combo, 2)

}
