% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_combine_rolling_sets.r
\name{mc_combine_rolling_sets}
\alias{mc_combine_rolling_sets}
\title{Combine a set of MCMC iterations created by 'mc_nimble_rolling()'}
\usage{
mc_combine_rolling_sets(model_dir)
}
\arguments{
\item{model_dir}{Folder in which the model output is saved.}
}
\value{
An \code{mcmc} object.
}
\description{
The function \code{\link[=mc_nimble_rolling]{mc_nimble_rolling()}} creates a series of files, each with a successive set of MCMC iterations (e.g., MCMC samples 1 through 1000, 1001 through 2000, etc.). This function combines them into a single \code{mcmc} object (see \code{\link[coda:as.mcmc]{coda::as.mcmc()}}). The files will be assumed ot be named like \code{chain_set_ZZZ.rds} where \code{ZZZ} is a number starting at 1 and going to the maximum number of sets, with zero-padding (e.g., \code{chain_set_001.rds}, \code{chain_set_002.rds}, etc.).
}
\examples{
man/examples/ex_nimble_rolling.r

}
\seealso{
\code{\link[=mc_nimble_rolling]{mc_nimble_rolling()}}
}
