% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_compile_chain.r
\name{mc_compile_chain}
\alias{mc_compile_chain}
\alias{mc_compile_chains}
\title{Combine a set of MCMC iterations created by 'mc_nimble_rolling()'}
\usage{
mc_compile_chain(model_dir)

mc_compile_chains(model_dirs)
}
\arguments{
\item{model_dir}{Folder in which the model output is saved.}

\item{model_dirs}{Folders in which output from multiple models is saved.}
}
\value{
An \code{mcmc} object.
}
\description{
\code{mc_compile_chain()} compiles a single chain from the files created by \code{\link[=mc_nimble_rolling]{mc_nimble_rolling()}} and deposited into one folder, and \code{mc_compile_chains()} compiles a series of chains from several folders, each populated with files created by \code{\link[=mc_nimble_rolling]{mc_nimble_rolling()}}.
}
\details{
The function \code{\link[=mc_nimble_rolling]{mc_nimble_rolling()}} creates a series of files, each with a successive set of MCMC iterations (e.g., MCMC samples 1 through 1000, 1001 through 2000, etc.). This function combines them into a single \code{mcmc} object (see \code{\link[coda:as.mcmc]{coda::as.mcmc()}}). The files will be assumed ot be named like \code{chain_set_ZZZ.rds} where \code{ZZZ} is a number starting at 1 and going to the maximum number of sets, with zero-padding (e.g., \code{chain_set_001.rds}, \code{chain_set_002.rds}, etc.).
}
\examples{
man/examples/ex_nimble_rolling.r

}
\seealso{
\code{\link[=mc_nimble_rolling]{mc_nimble_rolling()}}
}
