% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_n_chains.r
\name{mc_n_chains}
\alias{mc_n_chains}
\title{Number of chains in an mcmc object}
\usage{
mc_n_chains(mcmc)
}
\arguments{
\item{mcmc}{An object of class \code{mcmc} or \code{mcmc.list}, \emph{or} a \code{list}. If a \code{list}, the function searches down the first element to see if it can find an \code{mcmc} or \code{mcmc.list} object, then uses this if it can.}
}
\value{
Numeric.
}
\description{
Returns the number of chains in an \code{mcmc} list.
}
\examples{

data(mcmc)
mc_n_chains(mcmc)

}
