% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_n_iter.r
\name{mc_n_iter}
\alias{mc_n_iter}
\title{Number of iterations in an mcmc object}
\usage{
mc_n_iter(mcmc, total = FALSE)
}
\arguments{
\item{mcmc}{An object of class \code{mcmc} or \code{mcmc.list}, \emph{or} a \code{list}. If a \code{list}, the function searches down the first element to see if it can find an \code{mcmc} or \code{mcmc.list} object, then uses this if it can.}

\item{total}{Logical. If \code{TRUE}, then the value is the total number of iterations across all chains. If \code{FALSE} (default), then the value is the number in one chain.}
}
\value{
Numeric.
}
\description{
Returns the number of iterations in an \code{mcmc} list.
}
\examples{

data(mcmc)
mc_n_iter(mcmc)

}
