% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_check_nimble.r
\name{mc_nimble_check}
\alias{mc_nimble_check}
\title{Check a \code{nimble} model nodes for infinite or NaN likelihoods}
\usage{
mc_nimble_check(model, fail = FALSE, verbose = FALSE)
}
\arguments{
\item{model}{Output from \code{\link[nimble:nimbleModel]{nimble::nimbleModel()}}. The model must have all nodes initialized or \code{\link[nimble:calculate]{nimble::calculate()}} must have been run on the model to ensure all nodes could have non-missing values.}

\item{fail}{Logical: If \code{FALSE}, simply reports the node(s) that fail the checks. If \code{TRUE}, then reports the node and exits with an error.}

\item{verbose}{Logical: If \code{TRUE}, report progress.}
}
\value{
Invisibly returns \code{TRUE} if there are no nodes with infinite or \code{NaN} likelihood values. Otherwise, reports the node(s) that fail these checks.
}
\description{
Checks each node of a \code{nimble} model object created from \code{\link[nimble:nimbleModel]{nimble::nimbleModel()}} for likelihoods that are \code{NaN} or infinite.
}
\examples{
library(nimble)

n <- 100
x <- 1:n
beta0 <- 12
beta1 <- 3.7
linear_predictor <- beta0 + beta1 * x
sigma <- 42
y <- rnorm(n, linear_predictor, sd = sigma)

data <- list(y = y)
constants <- list(n = n, x = x)
inits <- list(beta0 = 0, beta1 = 0, sigma = 10)

code <- nimbleCode({
   beta0 ~ dnorm(0, sd = 100)
   beta1 ~ dnorm(0, sd = 100)
   sigma ~ dunif(0, 100)
   for (i in 1:n) y[i] ~ dnorm(beta0 + beta1 * x[i], sd = sigma)
})

model <- nimbleModel(
   code = code,
   constants = constants,
   data = data,
   inits = inits,
   check = TRUE
)
model$calculate()
mc_nimble_check(model)
}
