% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_rolling_nimble.r
\name{mc_nimble_rolling}
\alias{mc_nimble_rolling}
\title{Run a restartable 'nimble' model}
\usage{
mc_nimble_rolling(
  model_dir,
  code,
  config,
  compiled,
  niter_per_set,
  max_sets,
  inits = NULL,
  verbose = TRUE
)
}
\arguments{
\item{model_dir}{Folder in which the model output is to be saved.}

\item{code}{Output from \code{\link[nimble:nimbleCode]{nimble::nimbleCode()}}.}

\item{config}{Output from \code{\link[nimble:configureMCMC]{nimble::configureMCMC()}}.}

\item{compiled}{Output from \code{\link[nimble:compileNimble]{nimble::compileNimble()}}}

\item{niter_per_set}{Numeric (integer): Number of iterations per set. No thinning is done.}

\item{max_sets}{Numeric (integer): Total number of sets to do.}

\item{inits}{Either \code{NULL} (default) or a \code{list} of initialization values. OK to leave as \code{NULL} if not restarting.}

\item{verbose}{Logical: If \code{TRUE}, display progress.}
}
\value{
Nothing (saves files to disk).
}
\description{
This function runs a restartable \strong{nimble} model. Assuming it is starting "fresh" where no prior MCMC samples have been created, it first creates an output folder, generates MCMC samples, then saves them in a file that has the format \code{chain_set_XYZ.rds} where \code{XYZ} indicates the set number. The first set will thus be saved in the file \code{chain_set_001.rds}, the second in \code{chain_set_002.rds}, and so on. Successive sets are generated using the state of the sampler at the end of the previous set. This way, sets can be combined (e.g., using \code{\link[=mc_combine_rolling_sets]{mc_combine_rolling_sets()}}) to create a coherent chain. The state of the sampler is saved in files named like \code{.model_state_of_set_001.rds}, \code{.model_state_of_set_002.rds}, etc.
}
\details{
The function will also restart where it last left off if it is stopped or \strong{R} is stopped by loading the last \code{.model_state_of_set_XYZ.rds} file and continuing from there.

The function can only be run for a single chain at a time. If you wish to have more than one chain, create another folder for this chain and run the function in another instance of \strong{R}. Owing to this, WAIC cannot be calculated. The function also assumes a thinning rate of 1 (samples can be thinned post hoc).

Currently, restarting is supported for these samplers:
\itemize{
\item 'RW'
\item 'RW_block'
\item 'slice'
\item 'AF_slice'
\item 'RW_block_lkj_corr_cholesky' (Cholesky sampler from \code{\link[nimble:dlkj_corr_cholesky]{nimble::dlkj_corr_cholesky()}})).
Other samplers cannot be restarted :(
}
}
\examples{
library(nimble)

set.seed(1)

n <- 100
x <- 1:n
beta0 <- 12
beta1 <- 3.7
linear_predictor <- beta0 + beta1 * x
sigma <- 42
y <- rnorm(n, linear_predictor, sd = sigma)

data <- list(y = y)
constants <- list(n = n, x = x)
inits <- list(beta0 = 0, beta1 = 0, sigma = 1)

# nimble model
code <- nimbleCode({
   beta0 ~ dnorm(0, sd = 100)
   beta1 ~ dnorm(0, sd = 100)
   sigma ~ dunif(0, 100)
   for (i in 1:n) y[i] ~ dnorm(beta0 + beta1 * x[i], sd = sigma)
})

# build model
model <- nimbleModel(
   code = code,
   constants = constants,
   data = data,
   inits = inits,
   check = TRUE
)
model$calculate()
mc_nimble_check(model)

monitors <- c('beta0', 'beta1', 'sigma')
conf <- configureMCMC(
   model,
   monitors = monitors,
   print = FALSE,
   enableWAIC = FALSE
)
build <- buildMCMC(conf)
compiled <- compileNimble(model, build, showCompilerOutput = FALSE)

# MCMC samples to be saved here
model_dir <- tempdir()

# (using small number of iterations to show continuity between sets)
mc_nimble_rolling(
	model_dir = model_dir,
	code = code,
	config = config,
	compiled = compiled,
	niter_per_set = 20,
	max_sets = 3,
   inits = inits,
   verbose = TRUE
)

mcmc_files <- list.files(model_dir, pattern = 'chains_set_')
mcmc_files

# combine samples into a chain
chain <- mc_combine_rolling_sets(model_dir)

# trace plots
plot(chain[ , 'beta0'])
plot(chain[ , 'beta1'])
plot(chain[ , 'sigma'])

# run 2 more sets
mc_nimble_rolling(
	model_dir = model_dir,
	code = code,
	config = config,
	compiled = compiled,
	niter_per_set = 20,
	max_sets = 5,
   inits = inits,
   verbose = TRUE
)

# combine samples into a chain
chain <- mc_combine_rolling_sets(model_dir)

# trace plots
plot(chain[ , 'beta0'])
plot(chain[ , 'beta1'])
plot(chain[ , 'sigma'])

}
\seealso{
\code{\link[=mc_combine_rolling_sets]{mc_combine_rolling_sets()}} for combining the MCMC set files into a single MCMC chain.
}
