% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_rename.r
\name{mc_rename}
\alias{mc_rename}
\title{Rename a variable in ann MCMC object}
\usage{
mc_rename(mcmc, param_from, param_to, i = NULL, j = NULL, k = NULL, l = NULL)
}
\arguments{
\item{mcmc}{An object of class \code{mcmc} or \code{mcmc.list}, \emph{or} a \code{list}. If a \code{list}, the function searches down the first element to see if it can find an \code{mcmc} or \code{mcmc.list} object, then uses this if it can.}

\item{param_from}{Character: Name of the variable to rename.}

\item{param_to}{Character: New name.}

\item{i, j, k, l}{Indices used to specify variable names:
\itemize{
\item \code{i} refers to indices that occur immediately after the variable's "base" name. For example, if \code{param} is "\code{beta}", and \code{i = 0:2}, then the function will return \code{beta0}, \code{beta1}, and \code{beta2}.
\item \code{j} refers to the \emph{first} index in square brackets. For example, if \code{param} is \code{beta} and \code{j = 1:3}, then the output will be \code{beta[1]}, \code{beta[2]}, and \code{beta[3]}.
\item \code{k} refers to the \emph{second} index in square brackets. If \code{k} is defined (not \code{NULL}), then \code{j} must be defined. For example, if \code{param} is \code{beta}, \code{j = 1:2}, and \code{k = 1:3}, then the output will be like \code{beta[1, 1]}, \code{beta[2, 1]}, \code{beta[1, 2]}, \code{beta[2, 2]}, \code{beta[1, 3]}, and \code{beta[2, 3]}.
\item \code{l} refers to the \emph{third} index in square brackets. If you provide \code{l}, you must also define \code{j} and \code{k}. For example, if \code{param} is \code{beta}, \code{j = 1:2}, \code{k = 4:5}, and \code{l = 9:10}, then the output will be like \code{beta[1, 4, 9]}, \code{beta[2, 4, 9]}, \code{beta[1, 5, 9]}, \code{beta[2, 5, 9]}, \code{beta[1, 4, 10]}, \code{beta[2, 4, 10]}, and so on.
\item You can also define \code{i} when you define \code{j} alone, \code{j} and \code{k}, or \code{j}, \code{k}, and \code{l}. This will return patterns like \code{beta0[1]}, or like \code{beta0[1, 1]}, or like \code{beta0[1, 2, 3]}.
You must specify all of the indices that would appear in square brackets. For example, for the pattern \code{param[j, k, l]}, you need to define \code{j}, \code{k}, and \code{l}. You can't just define \code{k = 1:2}. You need to set \code{j} and \code{l} equal to numeric/integer vectors and/or logical values.
}}
}
\value{
An \code{mcmc.list} object or a \code{list}.
}
\description{
Sometimes we just goof up and name a variable something we later regret. Using this function, you can correct that error by renaming the variable in an MCMC object.
}
\examples{

data(mcmc)

# simple extraction
mc_param(mcmc)
mcmc_new <- mc_rename(
   mcmc, param_from = 'beta', param_to = 'gamma', j = 1:3
)
mc_param(mcmc_new)

}
