% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_resummarize.r
\name{mc_resummarize}
\alias{mc_resummarize}
\title{Re-calculate summaries of MCMC chains}
\usage{
mc_resummarize(mcmc, na.rm = FALSE)
}
\arguments{
\item{mcmc}{An object of class \code{mcmc} or \code{mcmc.list}, \emph{or} a \code{list}. If a \code{list}, the function searches down the first element to see if it can find an \code{mcmc} or \code{mcmc.list} object, then uses this if it can.}

\item{na.rm}{Logical: If \code{FALSE} (default), then summary statistics will be \code{NA} if any value in any iteration is \code{NA} (and it probably should be \code{NA}). However, if you set this to \code{TRUE}, then \code{NA} values are ignored.}
}
\value{
A \code{list} with the following structure:

\if{html}{\out{<div class="sourceCode">}}\preformatted{.$summary$chain1 <matrix>
.$summary$chain2 <matrix>
.$summary$chain3 <matrix>
.$summary$chain4 <matrix>
(etc.)
.$summary$all.chains <matrix>
}\if{html}{\out{</div>}}
}
\description{
This function re-calculates the summary statistics of MCMC chains.
}
\examples{

data(mcmc)
summ <- mc_resummarize(mcmc)
str(summ)

}
