% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_samples_summaries_summary.r
\name{mc_samples}
\alias{mc_samples}
\alias{mc_summaries}
\alias{mc_summary}
\title{Extracts a "samples" mcmc.list from a list, or the "summary" of an mcmc.list}
\usage{
mc_samples(mcmc, fail = TRUE)

mc_summaries(mcmc, fail = TRUE)

mc_summary(mcmc, fail = TRUE)
}
\arguments{
\item{mcmc}{An object of class \code{mcmc} or \code{mcmc.list}, \emph{or} a \code{list}. If a \code{list}, the function searches down the first element to see if it can find an \code{mcmc} or \code{mcmc.list} object, then uses this if it can.}

\item{fail}{Logical: If \code{TRUE} (default), and the object being searched for cannot be found, then fail with an error. If \code{FALSE}, then return \code{NULL}.}
}
\value{
An \code{mcmc.list} object or a \code{list}.
}
\description{
These functions extract the \code{samples} or \code{summary} part of a list object:
\itemize{
\item \code{mc_samples()}: Extracts a list of MCMC samples.
\item \code{mc_summaries()}: Extracts a list of MCMC summary matrices.
\item \code{mc_summary()}: Extracts the "all-chains" summary matrix.
}
}
\section{Functions}{
\itemize{
\item \code{mc_summaries()}: Extracts a "samples" list from a list

\item \code{mc_summary()}: Extracts a grand "summary" matrix from a list

}}
\examples{

data(mcmc)
str(mc_samples(mcmc))
str(mc_summaries(mcmc))
mc_summary(mcmc)

}
