% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_stack.r
\name{mc_stack}
\alias{mc_stack}
\title{Combine chains into one matrix}
\usage{
mc_stack(mcmc)
}
\arguments{
\item{mcmc}{An object of class \code{mcmc} or \code{mcmc.list}, \emph{or} a \code{list}. If a \code{list}, the function searches down the first element to see if it can find an \code{mcmc} or \code{mcmc.list} object, then uses this if it can.}
}
\value{
A \code{matrix}.
}
\description{
\code{\link[=rbind]{rbind()}}s chains onto one matrix.
}
\examples{

data(mcmc)

stacked <- mc_stack(mcmc)
dim(stacked)
head(stacked)
tail(stacked)

}
