% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_subset_chains.r
\name{mc_subset_chains}
\alias{mc_subset_chains}
\title{Select specific chains in an MCMC object}
\usage{
mc_subset_chains(mcmc, x, keep = TRUE, collapse = TRUE, na.rm = FALSE)
}
\arguments{
\item{mcmc}{An object of class \code{mcmc} or \code{mcmc.list}, \emph{or} a \code{list}. If a \code{list}, the function searches down the first element to see if it can find an \code{mcmc} or \code{mcmc.list} object, then uses this if it can.}

\item{x}{Numeric vector of chains.}

\item{keep}{Logical: If \code{TRUE} (default), chains named by \code{x} are retained. If \code{FALSE}, they are discarded.}

\item{collapse}{Logical: If \code{TRUE} (default), and just one chain is returned, then "collapse" the MCMC object (e.g., so it is structured just as \code{mcmc$samples} and \code{mcmc$summary}, versus the "uncollapsed" case where we would have \code{mcmc$samples$chain1}, and \code{mcmc$summary$chain1} and \code{mcmc$summary$all.chains}).}

\item{na.rm}{Logical: If \code{TRUE}, ignore \code{NA} values when re-summarizing retained chains. Default if \code{FALSE}.}
}
\value{
An \code{mcmc} matrix, an \code{mcmc.list}, or  a list of list, one of which is an \code{mcmc.list}.
}
\description{
Select specific chains in an MCMC object
}
\examples{

data(mcmc)

# retaining just one chain and collapsing
collapse <- mc_subset_chains(mcmc, 1)
str(collapse)

# retaining just one chain but not collapsing
no_collapse <- mc_subset_chains(mcmc, 1, collapse = FALSE)
str(no_collapse)

# discarding
disc <- mc_subset_chains(mcmc, 2, keep = FALSE)
str(disc)

}
