% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mch_stack.r
\name{mch_stack}
\alias{mch_stack}
\title{"Stack" multiple MCMC chains on one another}
\usage{
mch_stack(x)
}
\arguments{
\item{x}{An object of class \code{mcmc} or \code{mcmc.list}, or a \code{list}. If a \code{list}, the function searches down the first element to see if it can find an \code{mcmc} or \code{mcmc.list} object, then plots this if it can.}
}
\value{
Matrix.
}
\description{
This function takes a list of MCMC chains and "stacks" them onto one another. This can be useful for plotting.
}
