% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.r
\docType{data}
\name{mcmc}
\alias{mcmc}
\title{Example MCMC "list" object}
\format{
A \code{list}.
}
\description{
A \code{list} typical of output from \strong{rjags}, \strong{rstan}, or \strong{nimble}.
}
\keyword{MCMC}
