% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcHammer.r
\docType{package}
\name{mcmcHammer}
\alias{mcmcHammer-package}
\alias{mcmcHammer}
\title{mcmcHammer: Analysis of MCMC chains}
\description{
Functions for analyzing MCMC chains. Includes trace plots and density plots. This package automates extraction of specific parameters, including indexed coefficients with names such as "\code{alpha0}", "\code{alpha1}", or "\code{beta[1, 1]}", "\code{beta[1, 2]}", "\code{beta[1, 3]}", and similar, without the user needing to, for example, use \code{paste()} to create the variable names or to remember how many \code{beta}s there were.
}
\details{
Find a bug? Please report it and how it occurred on \href{https://github.com/adamlilith/mcmcHammer/issues}{GitHub}.
\subsection{Restartable \strong{nimble} models}{

\code{\link[=mc_nimble_check]{mc_nimble_check()}}: Check a \strong{nimble} model for nodes with \code{NaN} or infinite likelihoods.
\code{\link[=mc_nimble_rolling]{mc_nimble_rolling()}}: Run a restartable \strong{nimble} model.
\code{\link[=mc_compile_chain]{mc_compile_chain()}}: Combine a set of MCMC iterations created by \code{mc_nimble_rolling()}.
}

\subsection{Extraction and subsetting}{
\itemize{
\item \code{\link[=mc_extract]{mc_extract()}}: Mean, median, or lower or upper quantiles from the posterior distribution of a parameter.
\item \code{\link[=mc_samples]{mc_samples()}}: Extract an \code{mcmc.list} object with MCMC samples.
\item \code{\link[=mc_summaries]{mc_summaries()}}: Extract an "summary" list of matrices.
\item \code{\link[=mc_summary]{mc_summary()}}: Extract an "all=chains" summary matrix.
\item \code{\link[=mc_stack]{mc_stack()}}: Combine MCMC samples using \code{\link[=rbind]{rbind()}}.
\item \code{\link[=mc_subset]{mc_subset()}}: Subset MCMC chains by variable name and/or index.
\item \code{\link[=mc_subset_chains]{mc_subset_chains()}}: Select specific chains in an MCMC object.
}
}

\subsection{MCMC chain manipulation}{
\itemize{
\item \code{\link[=mc_cbind]{mc_cbind()}}: Combine side-to-side compatible \code{mcmc}, \code{mcmc.list}, or \code{list}s of \code{mcmc.list}s.
\item \code{\link[=mc_combine]{mc_combine()}}: Combine two or more MCMC objects and calculate new summary statistics.
\item \code{\link[=mc_rbind]{mc_rbind()}}: "Stack" MCMC chains into a single matrix.
\item \code{\link[=mc_rename]{mc_rename()}}: Rename one or more variables in an \code{mcmc} object.
}
}

\subsection{Helper functions:}{
\itemize{
\item \code{\link[=mc_n_chains]{mc_n_chains()}}: Number of chains.
\item \code{\link[=mc_n_iter]{mc_n_iter()}}: Number of iterations.
\item \code{\link[=mc_param]{mc_param()}}: Match variables names to MCMC columns.
\item \code{\link[=mc_resummarize]{mc_resummarize()}}: Per-chain and all-chains summary matrices.
}
}

\subsection{Example data}{
\itemize{
\item \code{mcmc}: An "MCMC \code{list}" similar to output from \strong{rstan}, \strong{rjags}, and \strong{nimble}.
}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/adamlilith/mcmcHammer}
  \item Report bugs at \url{https://github.com/adamlilith/mcmcHammer/issues}
}

}
\author{
Adam B. Smith
}
\keyword{internal}
