% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mh_stack.r
\name{mh_stack}
\alias{mh_stack}
\title{"Stack" multiple MCMC chains on one another}
\usage{
mh_stack(x, recursive = TRUE)
}
\arguments{
\item{x}{An object of class \code{mcmc} or \code{mcmc.list}, or a \code{list}. If a \code{list}, the function searches down the first element to see if it can find an \code{mcmc} or \code{mcmc.list} object, then plots this if it can.}

\item{recursive}{If \code{TRUE} (default), then recursively search \code{x} (assuming it is a list), for all elements that are objects of class \code{mcmc} and stack them. Note that if different list elements contain MCMC chains from different models (i.e., variables are not the same), this will fail.}
}
\value{
Matrix.
}
\description{
This function takes a list of MCMC chains and "stacks" them onto one another. This can be useful for plotting.
}
