% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emc_trace_dens.r
\name{emc_trace_dens}
\alias{emc_trace_dens}
\title{Trace plots and density plots}
\usage{
emc_trace_dens(
  x,
  var = NULL,
  i = NULL,
  j = NULL,
  k = NULL,
  l = NULL,
  trace = TRUE,
  loess = TRUE,
  nrows = 5,
  file = NULL,
  ...
)
}
\arguments{
\item{x}{An object of class \code{mcmc} or \code{mcmc.list}, or a \code{list}. If a \code{list}, the function searches down the first element to see if it can find an \code{mcmc} or \code{mcmc.list} object, then plots this if it can.}

\item{var}{Name of the variable(s) to plot. The outcome depends on the definitions of \code{i}, \code{j}, or \code{k}:
\itemize{
        \item \code{var = NULL}: Plot all monitored parameters.
        \item \code{var =} a character vector and \code{i}, \code{j}, and \code{k} are all \code{NULL}: Plot the variable(s) exactly as named.
        \item \code{var =} a character vector and \code{i} is a numeric vector: Plot all parameters with the pattern "\code{var*}" where \code{*} is \code{i}. For example: \code{beta1}, \code{beta2}, and \code{beta3}.
        \item \code{var =} a character vector and \code{i} is \code{TRUE}: Plot all parameters with the pattern \code{beta*} where \code{*} is 1 through the maximum value found in the \code{mcmc} object.
    \item \code{var =} a character vector and \code{j} is a numeric vector: Plot all parameters with the pattern "\code{beta[j]}".
    \item \code{var =} a character vector and \code{j} and \code{k} are a numeric vectors: Plot all parameters with the pattern "\code{beta[j, k]}".
    \item \code{var =} a character vector and \code{j}, \code{k}, and \code{l} are a numeric vectors: Plot all parameters with the pattern "\code{beta[j, k, l]}".
    \item \code{var =} a character vector and \code{i}, \code{j}, \code{k}, and \code{l} are a numeric vectors: Plot all parameters with the pattern "\code{beta*[j, k, l]}" where \code{*} is \code{i}.
    \item \code{var =} a character vector and any of \code{i}, \code{j}, \code{k}, and/or \code{l} are \code{TRUE}: Plot the given pattern up the the maximum value of \code{i}, \code{j}, \code{k}, and/or \code{l}.
}}

\item{i, j, kl}{These can be either \code{NULL} (default), integer vectors, or \code{TRUE}.}

\item{trace}{Logical. If \code{TRUE} (default), generate a trace plot.}

\item{loess}{Logical. If \code{TRUE} (default), draw a LOESS line for each chain in a trace plot.}

\item{nrows}{Number of rows of graphs to display per "page".}

\item{file}{Either \code{NULL} (default) or name of a file to which to save. Specifying a PDF file is especially helpful for cases where the}

\item{...}{Arguments to pass to \code{\link[ggplot2]{ggsave}}.}

\item{density}{Logical. If \code{TRUE} (default), generate a density plot.}
}
\value{
A \pkg{ggplot2} \code{ggplot} graphic object and (possibly) a file saved.
}
\description{
This function creates trace plots and/or density plots. It automates finding of indexed variable. These include, for example, variables like \code{beta1}, \code{beta2}, \code{beta3}, etc. or variables like \code{beta[1]}, \code{beta[2]}, \code{beta[1]}, etc.
}
