% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hammer_extract.r
\name{hammer_extract}
\alias{hammer_extract}
\title{Extract posterior mean, median, or quantiles from an mcmc or mcmc.list}
\usage{
hammer_extract(
  mcmc,
  param,
  i = NULL,
  j = NULL,
  k = NULL,
  l = NULL,
  stat = "mean"
)
}
\arguments{
\item{mcmc}{An object of class \code{mcmc} or \code{mcmc.list}, \emph{or} a \code{list}. If a \code{list}, the function searches down the first element to see if it can find an \code{mcmc} or \code{mcmc.list} object, then uses this if it can.}

\item{param}{Name of the variable(s). The outcome depends on the definitions of \code{i}, \code{j}, \code{k}, and \code{l}:
\itemize{
\item \code{param = NULL}: All variables in the MCMC object. An object must be supplied to the \code{mcmc} argument.
\item \verb{param =} a character vector and \code{i} is a numeric vector: Variables with the pattern '\verb{param*}' where \code{*} is \code{i}. For example: \code{beta1}, \code{beta2}, and \code{beta3}.
\item \verb{param =} a character vector and \code{i} is \code{TRUE}: Finds in the MCMC object all variables with the pattern '\verb{param*}' where \code{*} is numeric. You must supply \code{mcmc} in this case.
\item \verb{param =} a character vector and \code{j} is a numeric vector, but \code{k} is \code{NULL}: Variables with the pattern '\code{param[j]}'.
\item \verb{param =} a character vector and \code{j} is \code{TRUE}, but \code{k} is \code{NULL}: Variables with the pattern '\code{param[j]}'. You must supply \code{mcmc} in this case.
\item \verb{param =} a character vector and \code{j} and \code{k} are a numeric vectors but \code{l} is \code{NULL}: Variables with the pattern '\code{beta[j, k]}'.
\item \verb{param =} a character vector and \code{j} and/or \code{k} are \code{TRUE} but \code{l} is \code{NULL}: Variables with the pattern '\verb{beta[*, k]}', '\verb{beta[j, *]}', or '\verb{beta[*, *]}' (depending on values of \code{j} and \code{k}. You must supply \code{mcmc} in this case.
\item \verb{param =} a character vector and \code{j}, \code{k}, and \code{l} are a numeric vectors: Variables with the pattern '\code{beta[j, k, l]}'.
\item \verb{param =} a character vector and one or more of \code{j}, \code{k}, and \code{l} are a numeric vectors and/or \code{TRUE}: Variables with the pattern '\verb{beta[*, k, l]}', '\verb{beta[j, *, l]}', '\verb{beta[j, k, *]}', '\verb{beta[*, *, l]}', '\verb{beta[*, k, *]}',  '\verb{beta[j, *, *]}', or '\verb{beta[*, *, *]}'. You must supply \code{mcmc} in this case.
}}

\item{i, j, k, l}{Indices used to specify variable names. Please see the help for \code{\link[=hammer_param]{hammer_param()}}.}

\item{stat}{Character: Indicates which statistic to extract. Options are \code{'mean'} (default), \code{'median'}, \code{'sd'}, \code{'lower'}, '\verb{upper'}. The \code{lower} and \code{upper} options extract the lower and upper 2.5th and 97.5th quantiles, respectively. Partial matching is allowed.}
}
\value{
An \code{mcmc.list} object or a \code{list}.
}
\description{
Extracts posterior mean, median, standard deviation, or quantiles from an \code{mcmc} or \code{mcmc.list} object.
}
