% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hammer_rbind.r
\name{hammer_rbind}
\alias{hammer_rbind}
\title{"Stack" multiple MCMC chains on one another}
\usage{
hammer_rbind(mcmc, chains = FALSE)
}
\arguments{
\item{mcmc}{An object of class \verb{mcmc`` or }mcmc.list\verb{, or a `list}. If a \verb{list``, the function searches down the first element to see if it can find an }mcmc\code{or}mcmc.list` object, then plots this if it can.}

\item{chains}{Logical: If \code{TRUE}, then the output will have a column named \code{chain} with the chain number of each row.}
}
\value{
Matrix.
}
\description{
This function takes an \code{mcmc.list} and \code{rbind()}'s' them onto one another to create a single, large matrix.
}
