% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcHammer.r
\docType{package}
\name{mcmcHammer}
\alias{mcmcHammer-package}
\alias{mcmcHammer}
\title{mcmcHammer: Analysis of MCMC chains}
\description{
Functions for analyzing MCMC chains. Includes trace plots and density plots. This package automates extraction of specific parameters, including indexed coefficients with names such as "\code{alpha0}", "\code{alpha1}", or "\code{beta[1, 1]}", "\code{beta[1, 2]}", "\code{beta[1, 3]}", and similar, without the user needing to, for example, use \code{paste()} to create the variable names or to remember how many \code{beta}s there were.
}
\details{
Find a bug? Please report it and how it occurred on \href{https://github.com/adamlilith/mcmcHammer/issues}{GitHub}.
\subsection{Extraction and subsetting}{
\itemize{
\item \code{\link[=hammer_extract]{hammer_extract()}}: Mean, median, or lower or upper quantiles from the posterior distribution of a parameter.
\item \code{\link[=hammer_samples]{hammer_samples()}}: Extract an \code{mcmc.list} object with MCMC samples.
\item \code{\link[=hammer_summaries]{hammer_summaries()}}: Extract an "summary" list of matrices.
\item \code{\link[=hammer_summary]{hammer_summary()}}: Extract an "all=chains" summary matrix.
\item \code{\link[=hammer_subset]{hammer_subset()}}: Subset MCMC chains by variable name and/or index.
}
}

\subsection{MCMC chain manipulation}{
\itemize{
\item hammer_combine()]: Combine two or more MCMC objects and calculate new summary statistics.
\item \code{\link[=hammer_rbind]{hammer_rbind()}}: "Stack" MCMC chains into a single matrix.
}
}

\subsection{Helper functions:}{
\itemize{
\item \code{\link[=hammer_param]{hammer_param()}}: Match variables names to MCMC columns.
\item \code{\link[=hammer_resummarize]{hammer_resummarize()}}: Per-chain and all-chains summary matrices.
}
}

\subsection{Example data}{
\itemize{
\item \code{mcmc}: An "MCMC \code{list}" similar to output from \strong{rstan}, \strong{rjags}, and \strong{nimble}.
}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/adamlilith/mcmcHammer}
  \item Report bugs at \url{https://github.com/adamlilith/mcmcHammer/issues}
}

}
\author{
Adam B. Smith
}
\keyword{internal}
